requirements.txt
fitdecode==0.10.0
geographiclib==2.0
geopy==2.4.1
lxml==5.2.2
numpy==1.26.4
PySide6_Essentials==6.7.1
scipy==1.13.1
ruff==0.11.13
pyright==1.1.402
types-lxml[pyright]==2025.3.30
